+++ Riot/AppDelegate.m
@@ -3258,6 +3258,7 @@ NSString *const AppDelegateDidValidateEmailNotificationClientSecretKey = @"AppDe
                                                                                  }];
             }
 
+            __weak typeof(self) weakSelf = self;
             if (mxCall.isIncoming && isCallKitEnabled)
             {
                 // Let's CallKit display the system incoming call screen
@@ -3277,6 +3278,16 @@ NSString *const AppDelegateDidValidateEmailNotificationClientSecretKey = @"AppDe
 
                                                                                          NSLog(@"[AppDelegate] presentCallViewController");
                                                                                          [self presentCallViewController:NO completion:nil];
+                                                                                     }
+                                                                                     else if (call.state == MXCallStateEnded)
+                                                                                     {
+                                                                                         // Set call vc to nil to let our app handle new incoming calls even it wasn't killed by the system
+                                                                                         
+                                                                                         if (weakSelf)
+                                                                                         {
+                                                                                             typeof(self) self = weakSelf;
+                                                                                             [self dismissCallViewController:self->currentCallViewController completion:nil];
+                                                                                         }
                                                                                      }
                                                                                  }];
             }
@@ -3990,6 +4001,12 @@ NSString *const AppDelegateDidValidateEmailNotificationClientSecretKey = @"AppDe
                 [self dismissCallViewController:currentCallViewController completion:completion];
                 
             }];
+        }
+        else
+        {
+            // Release properly
+            [currentCallViewController destroy];
+            currentCallViewController = nil;
         }
     }
 }